+++ package.json
@@ -32,7 +32,6 @@
     "@radix-ui/react-checkbox": "latest",
     "@radix-ui/react-collapsible": "latest",
     "@radix-ui/react-context-menu": "latest",
-    "@radix-ui/react-dialog": "latest",
     "@radix-ui/react-dropdown-menu": "latest",
     "@radix-ui/react-hover-card": "latest",
     "@radix-ui/react-label": "latest",
@@ -44,7 +43,6 @@
     "@radix-ui/react-scroll-area": "latest",
     "@radix-ui/react-select": "latest",
     "@radix-ui/react-separator": "latest",
-    "@radix-ui/react-sheet": "^1.1.0",
     "@radix-ui/react-slider": "latest",
     "@radix-ui/react-slot": "latest",
     "@radix-ui/react-switch": "latest",
@@ -55,7 +53,7 @@
     "@radix-ui/react-tooltip": "latest",
     "@replit/vite-plugin-cartographer": "latest",
     "@replit/vite-plugin-runtime-error-modal": "latest",
-    "@shared/schema": "latest",
+    
     "@tailwindcss/typography": "latest",
     "@tanstack/react-query": "latest",
     "@tidbcloud/serverless": "latest",
